Fix: wrong vertical scaling for bars if not plot size was specifically set.
This was a sneaky one.
If user does not set a custom plot size, the library uses the window size.
This means that the plot takes up the whole window.
However, when setting the plot window size, the width was applied to both its width and height.
When calculating bar height, we use height relative to plot window height.
This means that bars got the wrong size because the plot height was set to the svg width initially..